[refactor] model 네이밍 위치 수정 #50
