+++ FindTown/FindTown/Domain/Entities/FilterModel.swift
@@ -1,0 +1,13 @@
+//
+//  FilterModel.swift
+//  FindTown
+//
+//  Created by 김성훈 on 2023/02/08.
+//
+
+import Foundation
+
+struct FilterModel {
+    var infra: String = ""
+    var traffic: [String] = []
+}
