+++ FindTown/FindTown/Domain/Entities/ReviewModel.swift
@@ -1,0 +1,14 @@
+//
+//  reviewModel.swift
+//  FindTown
+//
+//  Created by 김성훈 on 2023/02/08.
+//
+
+import Foundation
+
+struct ReviewModel {
+    let village: String
+    let period: String
+    let introduce: String
+}
