+++ FindTown/FindTown/Presentation/HomeScene/Flows/FilterSheetCoordinator.swift
@@ -13,13 +13,13 @@ final class FilterSheetCoordinator: FlowCoordinator {
     
     var presentationStyle: PresentationStyle
     var filterSheetType: FilterSheetType
-    var filterDataSource: TempFilterModel
+    var filterDataSource: FilterModel
     weak var navigationController: UINavigationController?
     
     init(
         presentationStyle: PresentationStyle,
         filterSheetType: FilterSheetType,
-        filterDataSource: TempFilterModel
+        filterDataSource: FilterModel
     ) {
         self.presentationStyle = presentationStyle
         self.filterSheetType = filterSheetType
@@ -35,7 +35,7 @@ final class FilterSheetCoordinator: FlowCoordinator {
 }
 
 extension FilterSheetCoordinator: FilterBottomSheetViewModelDelegate {
-    func dismiss(_ tempModel: TempFilterModel) {
+    func dismiss(_ tempModel: FilterModel) {
         guard let navigationController = navigationController else { return }
         if let homeVC = navigationController.viewControllers.first as? HomeViewController {
             homeVC.dismissBottomSheet(tempModel)
